origin double to long

+++ db/db_change.sql
@@ -1606,6 +1606,6 @@ ALTER TABLE `doctor_farms`
 ADD COLUMN `is_weak` SMALLINT(6) DEFAULT 1 COMMENT '弱仔数是否作为活仔数' AFTER `is_intelligent`;
 
 -- 猪 猪群原值 2017-12-20
-ALTER TABLE `doctor_pig_events` ADD COLUMN `origin` DOUBLE DEFAULT NULL COMMENT '原值' after `remark`;
-ALTER TABLE `doctor_group_events` ADD COLUMN `origin` DOUBLE DEFAULT NULL COMMENT '原值' after `remark`;
-ALTER TABLE `doctor_pigs` ADD COLUMN `origin` DOUBLE DEFAULT NULL COMMENT '原值' after `pig_code`;
+ALTER TABLE `doctor_pig_events` ADD COLUMN `origin` bigint(20) DEFAULT NULL COMMENT '原值' after `remark`;
+ALTER TABLE `doctor_group_events` ADD COLUMN `origin` bigint(20) DEFAULT NULL COMMENT '原值' after `remark`;
+ALTER TABLE `doctor_pigs` ADD COLUMN `origin` bigint(20) DEFAULT NULL COMMENT '原值' after `pig_code`;
